Fix CPU fractional values for Docker VMs.

+++ gns3server/schemas/controller/templates/docker_templates.py
@@ -49,7 +49,7 @@ class DockerTemplate(TemplateBase):
     extra_hosts: Optional[str] = Field("", description="Docker extra hosts (added to /etc/hosts)")
     extra_volumes: Optional[List] = Field([], description="Additional directories to make persistent")
     memory: Optional[int] = Field(0, description="Maximum amount of memory the container can use in MB")
-    cpus: Optional[int] = Field(0, description="Maximum amount of CPU resources the container can use")
+    cpus: Optional[float] = Field(0, description="Maximum amount of CPU resources the container can use")
     custom_adapters: Optional[List[CustomAdapter]] = Field(default_factory=list, description="Custom adapters")
